+++ Codes/shengyaqingfeng/.gitignore
@@ -1,1 +1,1 @@
-.vscode
+*.json
