--- Codes/shengyaqingfeng/1.两数之和.cpp
@@ -1,32 +1,0 @@
-/*
- * @lc app=leetcode.cn id=1 lang=cpp
- *
- * [1] 两数之和
- */
-#include <iostream>
-#include <vector>
-using namespace std;
-// @lc code=start
-class Solution {
-public:
-    vector<int> twoSum(vector<int>& nums, int target) {
-        vector<int> result(0);
-        for (int i = 0; i < nums.size() - 1; i++)
-        {
-            for (int j = i + 1; j < nums.size(); j++)
-            {
-                if (target == nums[i] + nums[j])
-                {
-                    result.push_back(i);
-                    result.push_back(j);
-                    return result;
-                }
-                
-            }
-            
-        }
-        return result;
-    }
-};
-// @lc code=end
-
