+++ Codes/shengyaqingfeng/README.md
@@ -1,6 +1,6 @@
 # LeetCode
 ## note
-1. 最熟悉的还是Java。但为了丢掉 `cpp` ，主要以 `cpp` 为主。无法独立实现才会切换 `Java`.
+1. 最熟悉的还是Java。但为了不丢掉 `cpp` ，所以主要以 `cpp` 为主。无法独立实现才会切换 `Java`.
 2. 坚决不看参考代码，不看评论。独立思考。实在没办法可以百度一下大体思路，然后独立实现。
 3. 计划 `2020-11-14` 开始，每周至少一题，一旦断更，下周补回。
 4. 允许搜索数据结构，搜索思路，搜索不熟悉的 `API` ，但代码一定独立完成。
@@ -10,5 +10,6 @@
 ## List of solved problems
 | date | Title | Running Time | Memory  | Note |
 |:----------:|:------|:------------:|:------------------:|:-----|
-| 2020-11-25 |[1. 两数之和](1.cpp)| | |            
-| 2020-11-25 |[2. 两数相加](2.cpp)| | |        
+| 2020-11-25 |[[1]Two Sum.cpp](leetcode/editor/cn/[1]Two Sum.cpp)| | |            
+| 2020-11-25 |[[2]Add Two Numbers.cpp](leetcode/editor/cn/[2]Add Two Numbers.cpp)| | |        
+| 2020-12-08 |[[3]Longest Substring Without Repeating Characters.cpp](leetcode/editor/cn/[3]Longest Substring Without Repeating Characters.cpp)| | |
